pyproject.toml
[tool.poetry]
name = "dynamic-pypi"
version = "0.5.0"
description = ""
authors = ["parnell <3028114+parnell@users.noreply.github.com>"]
readme = "README.md"

packages = [{ include = "dpypi" }]

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
pi-conf = "^0.7.6.4"
platformdirs = "^4.2.0"
pygithub = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
